fix(fgame): use uint32_t for md5_word_t to fix non-deterministic MD5 hashes (#928)
md5_word_t was typedef'd as size_t (64-bit on x86_64) instead of
uint32_t, causing out-of-bounds reads on X[8..15] in md5_process
and corrupt upper bits in T-constants. This made md5string return
random hashes.

+++ code/fgame/md5.h
@@ -50,6 +50,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 
 /*
  * This package supports both compile-time and run-time determination of CPU
@@ -62,7 +63,7 @@
  */
 
 typedef unsigned char md5_byte_t; /* 8-bit byte */
-typedef size_t md5_word_t; /* 32-bit word */
+typedef uint32_t md5_word_t; /* 32-bit word */
 
 /* Define the state of the MD5 Algorithm. */
 typedef struct md5_state_s {
